Add files via upload
Hiring process is the fundamental and the most important function of a company. We look at trends like number of positions, how many were hired/rejected, their salaries, etc. In this project I will be working for a MNC such as Google in analysing those data and run queries to get our answer to certain questions.